move and create resumableFileUploader() helper
