+++ projects/packages/videopress/src/client/lib/resumable-file-uploader/types.ts
@@ -1,0 +1,6 @@
+/**
+ * Internal dependencies
+ */
+import { VideoGuid, VideoId } from '../../block-editor/blocks/video/types';
+
+export type VideoMediaProps = { id: VideoId; guid: VideoGuid; src: string };
